refactor(server): cut duplicated reporter stub and excess commentary
Three test files had their own copy of the stub-and-restore reporter
boilerplate. Consolidates into stubApplicationErrorReporter in the shared
helpers both new tests and agent-stream already import.

Trims the reportHandlerFailure docblock from 15 lines to 4 and the
run-id guard comment from 3 to 1, keeping only what is not derivable
from the code. Sanitizer behavior is documented on the sanitizer.

No behavior change: same 63 steps pass.
